refactor: extract NSColor.controlAccentColor swizzle out of DataStore (#113)
Move the +controlAccentColor swizzle, the mer_currentTeamAccentColor
helper, the mer_invalidateAccentEverywhere broadcaster, and the
accentColorDidChange notification name to a new dedicated file
Meridian/App/AccentColorSwizzler.swift. DataStore is now scoped to
preference storage, with a one-line pointer comment to the new home.

Call sites:
- AppDelegate.applicationDidFinishLaunching: NSColor.installTeamAccentSwizzle()
  → AccentColorSwizzler.install().
- TeamAccentTests.testSwizzleMakesControlAccentColorReturnTeamAccent:
  same rename. Test still passes (full TeamAccentTests suite green).

Also drops the dead +mer_swizzledColorNamed extension method. It was
defined but never installed (only +controlAccentColor was swizzled);
the comment block above it explicitly documents that the colorNamed:
chase wasn't reliably working and was abandoned. Carrying dead code
into the new file would have been wrong.

No behavior change. The Once flag and NotificationCenter post are
preserved verbatim.
