+++ Meridian/Overall App/DataStore.swift
@@ -266,111 +266,8 @@ enum TeamAccent: String, Codable, CaseIterable {
 
 extension TeamAccent: JSONNameDecodable {}
 
-extension Notification.Name {
-    /// Posted when the user changes the team accent in Appearance settings.
-    /// PanelController observes this and triggers a redraw of the slider
-    /// fill + pin button tint.
-    static let accentColorDidChange = Notification.Name("com.tpak.meridian.accentColorDidChange")
-}
-
-// MARK: - controlAccentColor swizzle + invalidation
-//
-// Why we swizzle:
-// Without this, NSColor.controlAccentColor falls through to the
-// `Accent Color` colorset in Media.xcassets — a baked-in Aston Martin
-// green. So every NSPopUpButton selected indicator, NSSegmentedControl
-// segment fill, NSCheckbox check, NSToolbarItem tab icon, and focus
-// ring would stay green regardless of which team the user picks.
-// We swap the class-method getter at app launch so AppKit's internal
-// callers (and our own) get the live team color.
-//
-// Why a simple swizzle is not enough:
-// controlAccentColor is implemented as an NSDynamicSystemColor — a
-// dynamic system color whose resolution AppKit caches per
-// NSAppearance. Even after the swizzle returns a new color value,
-// AppKit reuses the previously-cached resolution for any surface that
-// has already been drawn. NSToolbarItem also caches its tinted
-// bitmap separately. Result: switching team A → team B in place does
-// nothing visible until some other event (tab nav, window
-// resize) forces a re-render.
-//
-// `NSApplication.mer_invalidateAccentEverywhere` performs the three
-// kinds of invalidation that, together, cover every visible surface:
-//   1. Re-set NSToolbarItem.image and NSTabViewItem.label/.image so
-//      the toolbar's tinted-bitmap cache is dropped.
-//   2. Recursive setNeedsDisplay on every view (and its CALayer) so
-//      layer-backed bitmaps repaint with the new accent.
-//   3. Toggle window.appearance through its current effective value
-//      so AppKit drops the NSDynamicSystemColor cache for that
-//      window's view tree without causing a light/dark flash.
-extension NSColor {
-    @objc class func mer_currentTeamAccentColor() -> NSColor {
-        DataStore.shared().teamAccent.accentColor
-    }
-
-    /// Intercepts `NSColor(named:)` lookups. When the name is the asset
-    /// catalog's accent entry ("AccentColor" / "Accent Color"), return
-    /// the live team accent. All other names fall through to the
-    /// original implementation.
-    ///
-    /// This is needed because AppKit's selected-toolbar-tab background
-    /// reads the asset catalog directly via `+colorNamed:` instead of
-    /// going through `+controlAccentColor`. Swizzling only the latter
-    /// left the tab background showing whatever was baked into
-    /// `Accent Color.colorset` (Aston Martin green) regardless of the
-    /// user's team selection.
-    @objc class func mer_swizzledColorNamed(_ name: String) -> NSColor? {
-        if name == "AccentColor" || name == "Accent Color" {
-            return DataStore.shared().teamAccent.accentColor
-        }
-        // After method_exchangeImplementations, this call lands on the
-        // original `+colorNamed:` because the implementations were
-        // swapped at runtime.
-        return mer_swizzledColorNamed(name)
-    }
-
-    /// Idempotent — guarded with a static flag so repeated calls are no-ops.
-    static func installTeamAccentSwizzle() {
-        struct Once { static var done = false }
-        guard !Once.done else { return }
-        Once.done = true
-
-        // Swizzle +controlAccentColor only. This covers NSPopUpButton,
-        // NSSegmentedControl, NSCheckbox, NSSlider, focus rings, etc.
-        // The selected-tab pill background is left to read the asset
-        // catalog (a neutral light gray) — chasing it through colorNamed:
-        // and private _selectionMaterialTintColor swizzles wasn't reliably
-        // working and the gray is unobtrusive enough that it doesn't fight
-        // the team color elsewhere.
-        let originalSel = #selector(getter: NSColor.controlAccentColor)
-        let customSel = #selector(NSColor.mer_currentTeamAccentColor)
-        guard let original = class_getClassMethod(NSColor.self, originalSel),
-              let custom = class_getClassMethod(NSColor.self, customSel) else {
-            return
-        }
-        method_exchangeImplementations(original, custom)
-    }
-}
-
-extension NSApplication {
-    /// Lets observers (PanelController, OneWindowController) repaint
-    /// the surfaces we explicitly tint with the team color: panel pin
-    /// button, slider chevrons + reset button, Preferences toolbar
-    /// icons (palette-baked SF Symbols).
-    ///
-    /// We do NOT attempt to invalidate AppKit's NSDynamicSystemColor
-    /// caches here. Every approach we tried (window.appearance
-    /// toggle, NSApp.deactivate/activate, tabStyle toggle, recursive
-    /// setNeedsDisplay) had at least one failure mode. macOS doesn't
-    /// expose a runtime API for changing controlAccentColor and
-    /// expecting AppKit's cached renderings to follow. The reliable
-    /// path is the one Apple takes for the system-wide accent change:
-    /// quit and relaunch. AppearanceViewController offers that as a
-    /// modal choice when the user picks a team.
-    func mer_invalidateAccentEverywhere() {
-        NotificationCenter.default.post(name: .accentColorDidChange, object: nil)
-    }
-}
+// The team-accent swizzle (NSColor.controlAccentColor + invalidation
+// notification) lives in App/AccentColorSwizzler.swift.
 
 // Indices match the popup item order in
 // AppearanceViewController.setupTimeFormatPopup(); 2/5/8 are disabled
